Start fresh, but not completely. Details follow:
Despite what it looks like, this is not another full rewrite. A lot of
clutter has been built up over time as I figure out what I'm doing,
since this is the only project I've led on this scale. I want to trim
out the fat and build the solution from scratch, pulling things in only
as I need them, and changing some of the architecture in the process
without dealing with debugging nightmare that comes with them.

This time, it will go much quicker. I'm generally happy with the way
things work now, but some naivity has permeated through the codebase and
it needs to get cleaned out. That's the goal of this rewrite.